% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{correlartic}
\alias{correlartic}
\title{Relação entre frequência de F1 e duração de vogal}
\format{
Um dataframe com 234 observações de três variáveis.
\describe{
  \item{vogal}{fator com sete níveis correspondentes às sete vogais
  tônicas do português brasileiro.}
  \item{dur}{vetor numérico, valor da duração da vogal medida em ms.}
  \item{F1}{vetor numérico, valor de F1 em Hz no ponto médio da vogal.}
}
}
\source{
Os dados foram gravados para o trabalho de colaboração de
Plinio A. Barbosa em 2014 para o Projeto Temático "Fronteiras e
assimetrias em fonologia e morfologia" coordenado por Filomena S.
Sandalo (DL/Unicamp).
}
\usage{
correlartic
}
\description{
Valores de duração em milissegundos (ms) e valores do primeiro
formante (F1) em Hertz (hz) de vogais orais de um informante masculino
do português paulista.
}
\details{
Um informante masculinos do interior de São Paulo com idade
entre 20 e 25 anos no momento da gravação leu um conjunto de frases
isoladas do tipo "Digo PALAVRA-CHAVE baixinho." Cada palavra-chave é
um vocábulo trissilábico paroxítono terminado em /a/. Palavras com
todas as vogais tônicas e pré-tôncias do português brasileiro
aparecerem nas palavras-chave. Este conjunto de dados contém apenas
a duração e F1 das vogais tônicas, medidas após a segmentação e
etiquetagem das vogais.
}
\examples{
data(correlartic)
summary(correlartic)
}
\keyword{datasets}
