% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{disfarce}
\alias{disfarce}
\title{Efeito de dois disfarces vocais sobre F1}
\format{
Um dataframe com 376 observações de duas variáveis.
\describe{
  \item{cond}{fator com 3 níveis, correspondentes ao disfarce leitura
  com lápis (\code{LAPIS}), imitação (\code{SILVI}) e fala habitual
  \code{HABIT}.}
  \item{F1}{vetor numérico, frequência do primeiro formante em Hertz
  das vogais orais nas três condições.}
}
}
\source{
Dados coletados em 2013 por Renata Passetti do Grupo de Estudos de
Prosódia da Fala da Universidade Estadual de Campinas para sua
pesquisa de Iniciação Científica, intitulada "Desviando da própria
fala: implicações para a verificação de locutor em situações de
disfarces vocais" (processo FAPESP 11/23054-3).
}
\usage{
disfarce
}
\description{
Dados de frequência em Hertz do primeiro formante (F1) das vogais
orais da leitura de um parágrafo em três condições de enunciação:
leitura habitual, leitura com lápis na boca, leitura imitando a fala
do Sílvio Santos.
}
\details{
A informante é paulista, cerca de 20 anos no momento da gravação e
leu um texto de cerca de 120 palavras de uma transcrição da fala do
apresentador de televisão Silvio Santos em três condições e três
vezes cada uma, escolhendo-se a melhor leitura.  Na primeira vez leu
de forma natural, com sua voz habitual; na segunda vez imitando o
apresentador de televisão, e, por fim, na terceira mordendo um lápis
na horizontal entre os dentes cerrados sem treinamento algum. As
vogais orais foram segmentadas e as frequências do primeiro formante
foram medidas em Hertz no ponto médio das mesmas.
}
\examples{
data(disfarce)
summary(disfarce)
}
\keyword{datasets}
