% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epentese}
\alias{epentese}
\title{Epêntese da vogal /i/ em português brasileiro}
\format{
Um dataframe com 71 observações de 9 variáveis.
\describe{
  \item{pal}{fator com 24 níveis, correspondentes a palavras contendo
  sequências /CiC/ e encontros consonantais /CC/ com possível epêntese
  de /i/.}
  \item{tipo}{fator com níveis \code{epe} (contextos em que pode
  acontecer epêntese) e \code{fon} (contextos em que o /i/ é
  fonológico).}
  \item{sil}{fator com níveis \code{pos} (sílaba pós-tônica) e
  \code{pre} (sílaba pré-tônica), correspondetes à tonicidade da
  sílaba em que a epêntese pode acontecer.}
  \item{cons}{fator com 12 níveis, correspondentes às consoantes nas
  sequências /CiC/ e /CC/ com possível epêntese.}
  \item{modo}{fator com níveis \code{FP} \code{NN} \code{PF} \code{PN}
  \code{PP}, correspondentes à classe das consoantes nas sequências
  /CiC/ e /CC/ com possível epêntese: fricativa (F), plosiva (P) e
  nasal (N).}
  \item{rep}{fator com 3 níveis, indica a repetição da sessão de
  gravação.}
  \item{vog}{fator com níveis \code{n} e \code{s}, indicando se a
  vogal /i/ está acusticamente ausente (\code{n}) ou presente
  (\code{s}) na palavra.}
  \item{dur}{vetor numérico, indicando a duração bruta da vogal /i/ em
  milissegundos quando ela está presente.}
  \item{norm}{vetor numérico, duração normalizada da vogal /i/ em
  z-\emph{scores} quando ela está presente.}
}
}
\source{
Dados coletados por Pablo Arantes em 2012 no Laboratório de Fonética
da Universidade Federal de Minas Gerais.
}
\usage{
epentese
}
\description{
Dados de ocorrência de epêntese vocálica em português brasileiro em
diferentes encontros consonantais.
}
\details{
Palavras-chave (coluna \code{pal}) foram inseridas em frases-veículo
do tipo "Digo \emph{palavra-chave} [bem devagar|para ele|lentamente]".
Foram gravadas três repetições de cada frase-veículo em uma única
sessão de gravação. A ordem das frases foi aleatorizada antes de cada
reapresentação da lista de estímulos à informante. As frases foram
apresentadas uma por vez na forma escrita. A informante é do sexo
feminino, natural de Recife-PE, com 40 anos na época da gravação.
}
\examples{
data(epentese)
summary(epentese)
}
\keyword{datasets}
