% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esforcovoc}
\alias{esforcovoc}
\title{Efeito da distância entre interlocutores sobre parâmetros acústicos}
\format{
Um dataframe com 143 observações de 4 variáveis.
\describe{
  \item{dist}{vetor numérico, valor da distância entre os
  interlocutores em metros.}
  \item{dur}{vetor numérico, valor da duração da vogal medida em
  milissegundos.}
  \item{f0med}{vetor numérico, valor da mediana da frequência
  fundamental em Hertz da vogal medida.}
  \item{enfespec}{vetor numérico, valor da ênfase espectral vogal
  medida em decibeis.}
}
}
\source{
Os dados foram coletados por Fernanda C. Grossi para sua Monografia
de curso intitulada "Esforço Vocal: uma análise de algumas variáveis
que o afetam", apresentada em 2014.
}
\usage{
esforcovoc
}
\description{
Efeito da distância entre interlocutores sobre a duração, frequência
fundamental (f0) e ênfase espectral em vogais de frases isoladas
produzidas por quatro informantes masculinos da variedade paulista do
português brasileiro.
}
\details{
Quatro informantes masculinos decoraram a frase curta "Compro muitas
frutas, carne e milho" e a usaram para responder à  pergunta "O que
você compra quando vai à feira?" feita por um interlocutor (o
experimentador) localizado a 5 distâncias distintas, de tal forma que
a frase fosse ouvida claramente. As distâncias foram baseadas em
trabalho prévio de Lindh e Eriksson (2007): 0,30 m, 1,5 m, 3 m, 25 m
e 40 m. Todos os informantes são do interior de São Paulo com idade
entre 20 e 25 anos no momento da gravação. As vogais das
palavras-chave foram segmentadas e etiquetadas. Um script elaborado
por Plinio A. Barbosa em 2014 permitiu a extração das variáveis
dependentes presentes no conjunto de dados.
}
\examples{
data(esforcovoc)
summary(esforcovoc)
}
\references{
LINDH, J.; ERIKSSON, A. Robustness of Long Time Measures of
Fundamental Frequency. \emph{In}: 8TH ANNUAL CONFERENCE OF THE
INTERNATIONAL SPEECH COMMUNICATION ASSOCIATION, 2007, Antwerp,
Belgium. \emph{Proceedings of the 8h Annual Conference of the
International Speech Communication Association}. Antwerp, Belgium,
2007. p. 2025–2028.
}
\keyword{datasets}
