% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{estilos}
\alias{estilos}
\title{Taxas de elocução e articulação em dois estilos de elocução}
\format{
Um dataframe com 161 observações de 3 variáveis.
\describe{
  \item{estilo}{fator com 2 níveis, indicando se a fala é lida
  (\code{RE}) ou narrada (\code{ST}).}
  \item{taxaeloc}{vetor numérico, valor da taxa de elocução de um
  trecho de fala pré-definido.}
  \item{taxaartic}{vetor numérico, valor da taxa de articulação de um
  trecho de fala pré-definido.}
}
}
\usage{
estilos
}
\description{
Valores de taxa de elocucão e taxa de articulação em sílabas fonéticas
por segundo nas falas lida e narrada em 10 informantes da variadade
paulista do português brasileiro.
}
\details{
Dez informantes leram um texto de cerca de 1600 palavras sobre a
origem dos pastéis de Belém e em seguida o narraram. Trechos
aleatórios entre 10 e 20 segundos foram extraídos com a anotação das
sílabas fonéticas nos dois estilos de elocução. Um script do Praat
calculou a taxa de elocução desses trechos, o que inclui as paulas
silenciosas. O mesmo script suprime as sílabas fonéticas pré-pausais
e correspondentes a picos de duração local para calcular uma medida
mais conservadora da taxa de articulação. Gera-se assim um arquivo de
texto que contém a identificação dos informantes e do estilo de cada
trecho. Todos os informantes são do interior de São Paulo com idade
entre 20 e 35 anos no momento da gravação. Seus dados foram agrupados
porque se comportam da mesma maneira para os dois tipos de taxa.
}
\examples{
data(estilos)
summary(estilos)
}
\keyword{datasets}
