% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{f0_modal}
\alias{f0_modal}
\title{F0 de assertivas e interrogativas em português brasileiro}
\format{
Um dataframe com 288 observações de 8 variáveis.
\describe{
	\item{speaker}{fator com 4 níveis, correspondentes aos falantes
	presentes na amostra, dois femininos (\code{fem.1} e \code{fem.2}) e
	dois masculinos (\code{masc.1} e \code{masc.2}).}
	\item{modal}{fator com 2 níveis, correspondentes às modalidades de
	frase, assertiva \code{asser} e interrogativa \code{inter}.}
	\item{position}{fator com 6 níveis, correspondentes ao índice das
	sílabas que formam o contorno.}
	\item{label}{fator com 6 níveis, correspondentes às sílabas das
	palavras presentes no contorno.}
	\item{sample}{fator com 36 níveis, índice da amostra normalizada de
	f0 no contorno.}
	\item{f0}{vetor numérico, valor de f0 (em Hertz) de cada amostra.}
	\item{step}{vetor numérico, diferença de tempo (em segundos) entre
	as amostras normalizadas.}
	\item{time}{vetor numérico, instante de tempo real (em segundos) em
	que a amostra normalizada foi coletada.}
}
}
\usage{
f0_modal
}
\description{
Valores de f0 em contornos normalizados temporalmente nas modalidades
assertiva e interrogativas em português brasileiro.
}
\details{
Contornos de f0 da frase "Marina cantava" enunciadas em duas
modalidades (assertiva e interrogativa) por quatro falantes nativos
do português brasileiro.

Os contornos de f0 foram extraídos das amostras de fala
disponibilizadas em conjunto com o texto da dissertação de Mestrado
de Luma Miranda (2015).

As curvas de f0 foram extraídas usando o programa de análise acústica
Praat e posteriormente normalizadas temporalmente usando o script
\code{time-normalized-f0} (Arantes, 2021).

A normalização temporal foi aplicada tendo como unidade de referência
a sílaba. Valores de f0 foram amostrados em seis pontos
isocronicamente espaçados. Cada contorno completo é composto por um
total de 36 amostras de f0. A curva de f0 bruta passou por suavização
usando um parâmetro de largura de banda de 10 Hz.
}
\examples{
data(f0_modal)
summary(f0_modal)
}
\references{
ARANTES, P. \emph{Time-normalized-f0: Praat script to perform
time-normalization of F0 contours}. Versão v2.0. Zenodo, 2021.
Disponível em: \url{https://zenodo.org/record/4489748}.

MIRANDA, L. da S. \emph{Análise da entoação do português do Brasil
segundo o modelo IPO}. 2015. 161 f. Dissertação de Mestrado -
Universidade Federal do Rio de Janeiro, Rio de Janeiro, 2015.
}
\keyword{datasets}
