% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{f0_styles}
\alias{f0_styles}
\title{F0 em três estilos de fala}
\format{
Um dataframe com 3.865.793 observações de 5 variáveis.
\describe{
	\item{lang}{fator com 5 níveis, correspondentes aos códigos das sete
	línguas de acordo com o padrão ISO 639 -- \code{deu} (alemão),
	\code{eng} (inglês), \code{est} (estoniano), \code{fra} (francês)
	\code{ita} (italiano), \code{por} (português) e \code{swe} sueco}
	\item{style}{fator com três níveis, correspondentes aos três estilos
	de elocução}
	\item{speaker}{fator com 10 níveis, correspondentes aos dez
	informantes de cada língua, cinco do sexo feminino (\code{f1} a
	\code{f5}) e cinco do sexo masculino (\code{m1} a \code{m5})}
	\item{time}{vetor numérico, instante de tempo na gravação em segundos
	em que cada valor de f0 foi coletado.}
	\item{f0}{vetor numérico, valor de f0, medido em Hertz.}
	}
}
\source{
Os contornos de f0 das gravações foram extraídos usando o programa de
análise acústica Praat por meio de scripts especializados (Arantes,
2019, 2021) e posteriormente verificados manualmente por até quatro
pessoas treinadas para identificar possíveis erros de extração. As
seguintes pessoas trabalharam para a extração em seus projetos de
pesquisa, orientadas por Pablo Arantes (DL/UFSCar): Maria Érica
Linhares (FAPESP processo 2014/21161-5), Suska Gutzeit
(PIBIC-CNPq/UFSCar edital 2014-2015) e Isabela Silveira (FAPESP
processo 2016/16544-8). Análises dos dados de f0 presentes neste
conjunto de dados são reportadas em Arantes (2021) e Arantes e
Linhares (2017).
}
\usage{
f0_styles
}
\description{
Valores de f0 de fala produzida em três estilos de elocução em sete
línguas.
}
\details{
Os contornos de f0 contidos neste conjunto de dados fazem parte de um
banco de dados de gravações chamado "A typology for word stress and
speech rhythm based on acoustic and perceptual considerations", reunido
por Anders Eriksson (Universidade de Estocolmo, Suécia) e projetado
para estudar o acento lexical em várias línguas. No momento em que o
conjunto de dados foi gerado, as línguas presentes no corpus eram
alemão, estoniano, francês, inglês britânico, italiano, português
brasileiro e sueco.

O corpus foi elaborado para elicitar três diferentes estilos de
elocução: fala espontânea, leitura de frases e leitura de palavras. A
fala espontânea foi obtida por meio de entrevistas semidirigidas
informais com os participantes, conduzidas por um falante nativo que
trabalhava no projeto. Essas gravações foram então transcritas
ortograficamente e usadas para produzir o material para os outros dois
estilos de fala.

Frases contendo palavras-alvo adequadas para experimentos de acento
lexical foram selecionadas das gravações das entrevistas, a partir de
trechos de fala fluente que não apresentavam erros de fala. Na etapa
seguinte, os falantes foram chamados novamente e solicitados a ler as
frases e palavras que haviam produzido em sua fala espontânea. Dessa
forma, foi possível obter o mesmo conteúdo linguístico nos três estilos
de fala.

Os participantes foram selecionados de modo a minimizar variações
devido a diferenças linguísticas regionais e de idade. Todos os
participantes recrutados falavam um padrão regional bem definido. A
variação etária dos falantes foi mantida dentro de margens estreitas
para todas as línguas analisadas. Foram recrutados cinco falantes do
sexo feminino e cinco do sexo masculino de cada língua.
}
\examples{
data(f0_styles)
summary(f0_styles)
}
\references{
ARANTES, P. \emph{Better_f0: Praat script for better f0 extraction}.
Versão v1.0. Zenodo, 2019. Disponível em:
\url{https://zenodo.org/record/3470107}.

ARANTES, P. \emph{Effects of speaking style on the shape of
fundamental frequency distributions}. Revista da ABRALIN, p. 1–39,
2021.

ARANTES, P. \emph{F0-outliers: Identify possible extraction errors in
Pitch objects}. Versão v1.0. Zenodo, 2021. Disponível em:
\url{https://zenodo.org/record/5083453}.

ARANTES, P.; LINHARES, M. E. N. Efeito da língua, estilo de elocução
e sexo do falante sobre medidas globais da frequência fundamental.
\emph{Letras de Hoje}, v. 52, n. 1, p. 26–39, 2017.

WIKIMEDIA FOUNDATION. \emph{List of ISO 639 language codes}.
Wikipedia. Disponível em:
\url{https://en.wikipedia.org/wiki/List_of_ISO_639_language_codes}.
Acesso em: 03 dez. 2024.
}
\keyword{datasets}
