% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{fmtn}
\alias{fmtn}
\title{Round a number}
\usage{
fmtn(x, dec = ",")
}
\arguments{
\item{x}{a numeric value.}

\item{dec}{a character representing the decimal separator to be used.}
}
\value{
Return a rounded number following American Psychological Association's
guidelines for presentation on reports.

Numbers are rounded as follows:

\tabular{llll}{
  Range     \tab Decimal places         \tab Input    \tab Output\cr
  > 100     \tab whole number           \tab 103.862  \tab 103\cr
  10-100    \tab 1                      \tab 48.69    \tab 48.7\cr
  0.1-10    \tab 2                      \tab 7.724    \tab 7.72\cr
  0.001-0.1 \tab 3                      \tab 0.0238   \tab 0.024\cr
  < 0.001   \tab as needed for non-zero \tab 0.000573 \tab 0.0006
}
}
\description{
Round a number for presentation purposes.
}
\examples{
fmtn(103.862)
fmtn(48.69)
fmtn(7.724)
fmtn(0.0238)
fmtn(0.000573)
}
