% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{fmtp}
\alias{fmtp}
\title{Format p-values}
\usage{
fmtp(x, dec = ",")
}
\arguments{
\item{x}{a numeric value representing a \emph{p}-value.}

\item{dec}{a character representing the decimal separator to be used.}
}
\value{
Return a character vector containing a formatted \emph{p}-value
following the American Psychological Association's guidelines.

Exact \emph{p}-values between 1 and 0.001 are rounded as follows:
\itemize{
  \item \emph{p} > 0.1: 2 decimal places
  \item \emph{p} >=  0.001: 3 decimal places
  \item \emph{p} < 0.001: "p < 0.001"
}
}
\description{
Format \emph{p}-values for printing.
}
\examples{
fmtp(0.367495)
fmtp(0.06942)
fmtp(0.0008503)
}
\references{
FIELD, Andy; HOLE, Graham. How to Design and Report Experiments. London: Sage
Publications, 2003.
}
