% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{formants}
\alias{formants}
\title{Formantes de vogais orais do português brasileiro}
\format{
Um dataframe com 1080 observações de 8 variáveis.
\describe{
 \item{\code{vogal}}{fator com 9 níveis, seis vogais pré-tônicas e
 três pós-tônicas.}
 \item{\code{arquivo}}{fator com 90 níveis, palavra-chave e
 repetição.}
 \item{\code{f1_m}}{vetor numérico, frequência em Hertz de F1
 extraído pelo método manual.}
 \item{\code{f2_m}}{vetor numérico, frequência em Hertz de F2
 extraído pelo método manual.}
 \item{\code{f3_m}}{vetor numérico, frequência em Hertz de F3
 extraído pelo método manual.}
 \item{\code{f1_a}}{vetor numérico, frequência em Hertz de F1
 extraído pelo método automático.}
 \item{\code{f2_a}}{vetor numérico, frequência em Hertz de F2
 extraído pelo método automático.}
 \item{\code{f3_a}}{vetor numérico, frequência em Hertz de F3
 extraído pelo método automático.}
}
}
\source{
Os dados acústicos foram coletados por Pablo Arantes no Instituto de
Estudos da Linguagem da Universidade Federal de Campinas na sala do
Grupo de Estudos de Prosódia da Fala, em 2008.
}
\usage{
formants
}
\description{
Valores em Hertz do primeiro, segundo e terceiro formantes (F1, F2 e
F3) de seis vogais orais tônicas e três vogais orais pós-tônicas do
português brasileiro. Os dados foram obtidos de modo a permitir a
comparação de desempenho de um humano (um foneticista treinado) e de
um algoritmo na tarefa de extrair formantes. Para cada vogal há,
portanto, seis valores, dois para cada formante.
}
\details{
As vogais foram inseridas em palavras-chave com estrutura /pV1pV2/,
em que V1 é /a eh e i o u/ e V2 é /A I U/. A palavra-chave foi
inserida em uma frase-veículo do tipo "Digo \emph{palavra-chave}
baixinho". A combinação de V1 e V2 gera 18 palavras-chave. Foram
gravadas cinco repetições de cada frase-veículo em uma única sessão
de gravação. A ordem das frases foi aleatorizada antes de cada
reapresentação da lista de estímulos ao informante. As frases foram
apresentadas uma por vez na forma escrita. O informante é do sexo
masculino, natural de Piracicaba-SP, com 39 anos na época da gravação.

A extração manual foi feita em torno do ponto médio da duração
acústica de cada vogal, comparando o espectro de LPC com o espectro
de FFT. O analista podia determinar quantos formantes deveriam ser
encontrados e a frequência máxima da faixa em que os formantes seriam
procurados.

A extração automática foi feita por um script do programa de análise
acústica Praat, programado por Leonardo Oliveira. O script busca a
combinação do número de formantes com a frequência máxima da faixa de
busca que minimiza a variabilidade do contorno de formantes ao longo
de toda a duração de uma vogal.
}
\examples{
data(formants)
summary(formants)
}
\references{
OLIVEIRA, Leonardo; ARANTES, Pablo. An algorithm for token-by-token
linear predictive coding analysis of oral monophthongs. \emph{The
Journal of the Acoustical Society of America}, v. 125, n. 4, p. 2529,
2009.
}
\keyword{datasets}
