% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{forster}
\alias{forster}
\title{Tempo de reação em experimentos de processamento lexical.}
\format{
Um dataframe com 30 observações de 5 variáveis.
\describe{
 \item{word}{fator com 30 níveis, palavras da língua inglesa
  testadas nos dois experimentos.}
 \item{cat}{fator com 2 níveis, faixa de frequência de ocorrência:
 alta (\code{high}) e baixa (\code{low}).}
 \item{rt_naming}{vetor numérico, tempo de reação médio na tarefa de
  nomeação em milissegundos.}
 \item{rt_lexdec}{vetor numérico, tempo de reação médio na tarefa de
  decisão lexical em milissegundos.}
 \item{freq_abs}{vetor numérico, frequência de ocorrência absoluta
  da palavra.}
}
}
\usage{
forster
}
\description{
Tempos de reação em milissegundos (ms) em um experimento de nomeação
(\emph{naming}) e em um de decisão lexical (\emph{lexical decision})
em língua inglesa com palavras de alta e baixa frequência de
ocorrência.
}
\details{
Os dados de tempo de reação foram tirados de Forster e Chambers
(1973). Os valores de tempo de reação para cada palavra representam,
em cada tarefa, a média de 20 participantes.

A frquência absoluta das palavras não é fornecida no artigo citado.
Os dados de frequência listados neste dataframe foram obtidos por
meio de consulta ao \href{http://corpus.byu.edu/coca/}{Corpus of Contemporary American English},
compilado por Mark Davies, da Brigham Young University. O
\emph{corpus} é composto por textos escritos e orais de diversas
proveniências (fala, ficção, revistas, jornais e textos acadêmicos),
produzidos entre 1990 e 2012, totalizando 450 milhões de ocorrências.
}
\examples{
data(forster)
summary(forster)
}
\references{
FORSTER, K. I.; CHAMBERS, S. M. Lexical Access and Naming Time.
\emph{Journal of Verbal Learning and Verbal Behavior}, v. 12, p.
627–635, 1973.
}
\keyword{datasets}
