% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intel}
\alias{intel}
\title{Inteligibilidade de aprendizes brasileiros em inglês L2}
\format{
Um dataframe com 120 observações de 6 variáveis.
\describe{
	\item{Aluno}{fator com 15 níveis, correspondentes ao código
	associado a cada aluno participante.}
	\item{Nivel}{fator com 2 níveis, correspondentes ao ano no ensino
	fundamental cursado pelos alunos participantes.}
	\item{Juiz}{fator com 4 níveis, correspondentes ao código associado
	aos quatro juízes falantes nativos de inglês.}
	\item{Palavra}{fator com 8 níveis, correspondentes às palavras
	transcritas e julgadas pelos juízes.}
	\item{Transcricao}{fator com dois níveis, correspondentes à condição
	da transcição da palavras produzida pelo aprendiz e transcrita pelo
	juiz nativo, \code{errado} e \code{certo}.}
	\item{Dificuldade}{vetor numérico, notas de 1 a 5, correspondentes
	ao julgamento de dificuldade de inteligibilidade da produção oral do
	aprendiz feito pelo juiz nativo.}
}
}
\source{
MENDES, M. L. \emph{Avaliação de inteligibilidade, compreensibilidade
e grau de sotaque estrangeiro do inglês-L2 de alunos brasileiros do
6º e 9º anos do ensino fundamental em sistema bilíngue de ensino}.
2024. Dissertação de Mestrado - Universidade Federal do Ceará,
Fortaleza, 2024.
}
\usage{
intel
}
\description{
Inteligibilidade de produções orais em inglês L2 de alunos dos 6º e
9º anos do ensino fundamental julgadas por falantes nativos de inglês
britânico.
}
\details{
Os dados resultam de pesquisa que teve como participantes alunos do
6º e 9º anos do ensino fundamental (Nível) de uma escola particular
que usa o sistema bilíngue de ensino (aulas de inglês 1 vez por
semana em estilo CLIL - Content and Language Integrated Learning),
que foram gravados contando uma história com base em figuras. Depois,
algumas palavras foram recortadas do áudio para que 4 juízes falantes
nativos de inglês britânico ouvissem e tentassem transcrever as
palavras. As transcrições foram marcadas como certas ou erradas. Além
disso, para cada palavra, o juiz tinha que marcar numa escala de 1 a
5 o grau de dificuldade que teve em compreender a palavra, sendo 1
muito fácil de compreender e 5 muito difícil de compreender.
Transcrever certo ou errado é uma medida da inteligibilidade do aluno
(neste caso, inteligibilidade local, por serem palavras isoladas), e
o grau de dificuldade que o juiz teve é o grau de compreensibilidade
do aluno, segundo as propostas de Derwing e Munro.
}
\examples{
data(intel)
summary(intel)
}
\keyword{datasets}
