% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{l2lexdec}
\alias{l2lexdec}
\title{Decisão lexical em segunda língua}
\format{
Um dataframe com 20 observations de 2 variáveis.
\describe{
  \item{length}{vetor numérico, número de letras na palavra}
  \item{rt}{vetor numérico, tempo de reação em milissegundos}
}
}
\source{
Conjunto de dados retirado do exemplo discutido por Gries
(2009, p. 151).
}
\usage{
l2lexdec
}
\description{
Dados de tempo de reação em tarefa de decisão lexical de aprendizes
de segunda língua (L2) para palavras com diferente número de letras.
}
\examples{
data(l2lexdec)
summary(l2lexdec)
}
\references{
GRIES, S. Th. \emph{Statistics for Linguistics with R: A Practical
Introduction}. Amsterdam: Mouton de Gruyter, 2010.
}
\keyword{datasets}
