% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{langpop99}
\alias{langpop99}
\title{Número de falantes por língua}
\format{
Um dataframe com 6783 observações de 4 variáveis.
\describe{
 \item{speakers}{vetor numérico, número de falantes nativos da
 língua.}
 \item{country}{fator com 201 níveis, nome do país onde a língua é
 falada.}
 \item{language}{fator com 6661 níveis, nome da língua.}
 \item{code}{fator com 6783 níveis, código identificador da língua
 segundo a norma ISO 639-3.}
}
}
\source{
O arquivo contendo a lista foi baixado do site do
\href{http://www.sil.org/ftp/ethnolog13/langpop99.zip}{SIL} em 23 de
agosto de 2013.
}
\usage{
langpop99
}
\description{
Lista do número de falantes por língua preparada pelo catálogo
Ethnologue em 1999.
}
\details{
A publicação Ethnologue é uma iniciativa ligada ao Summer Institute
of Linguistics (SIL) e se propõe a ser um catálogo periodicamente
atualizado de todas as línguas vivas conhecidas. Os números de
falantes da presente lista são uma atualização feita em 1999 dos
dados publicados na 13ª edição do Ethnologue, lançada em 1996. As
línguas com zero falantes deixaram de ser correntemente usadas a
partir de 1951, quando a primeira edição do catálogo foi lançada.
}
\examples{
data(langpop99)
summary(langpop99)
}
\references{
\url{http://www.ethnologue.com/about}
}
\keyword{datasets}
