% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingdata-package.R
\docType{package}
\name{lingdata-package}
\alias{lingdata-package}
\title{lingdata: Conjuntos de Dados Linguisticos}
\description{
Coleção de dados de natureza linguística destinada ao ensino de estatística para estudantes e pesquisadores das áreas de Letras e Linguística.
}
\author{
\strong{Maintainer}: Pablo Arantes \email{pabloarantes@ufscar.br}

Authors:
\itemize{
  \item Plinio Barbosa \email{pabarbosa.unicampbr@gmail.com}
  \item Ronaldo Mangueira Lima Júnior \email{ronaldo.limajr@gmail.com}
}

}
\keyword{internal}
