% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingdata.R
\name{lingdata}
\alias{lingdata}
\title{lingdata: Conjuntos de dados de natureza linguística.}
\description{
Coleção de dados de natureza linguística destinada ao ensino de estatística
para estudantes e pesquisadores das áreas de Letras e Linguística.
}
