% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{moda}
\alias{moda}
\title{Sample Mode}
\usage{
moda(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric, character or logical vector.}

\item{na.rm}{a logical value indicating whether NA values should be stripped
before the computation proceeds.}
}
\value{
A vector containing the most common value in the input vector. If the
sample is plurimodal, all modal values will appear. Input and output values
are of the same type.
}
\description{
Compute the sample mode of a vector.
}
\examples{
set.seed(321)
data <- sample.int(n = 10, size = 50, replace = TRUE)
table(data)
moda(data)
moda(letters[data])
moda(c(TRUE, TRUE, TRUE, FALSE, TRUE, FALSE))
}
