% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motherese}
\alias{motherese}
\title{Tamanho de enunciados de adultos dirigidos a crianças}
\format{
Um dataframe com 200 observações de 2 variáveis.
\describe{
 \item{\code{group}}{fator com níveis \code{A}, crianças de 18 meses
 e \code{B},crianças de 36 meses.}
 \item{\code{size}}{vetor numérico, tamanho dos enunciados medido em
 morfemas.}
}
}
\source{
WOODS, A.; FLATCHER, P.; HUGHES, A. \emph{Statistics in language
studies}. Cambridge: Cambridge University Press, 1986.
}
\usage{
motherese
}
\description{
Tamanho de enunciados de adultos dirigidos a crianças de duas idades
diferentes.
}
\examples{
data(motherese)
summary(motherese)
}
\keyword{datasets}
