% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quantidade}
\alias{quantidade}
\title{Duração de vogais simples e duplas em PB}
\format{
Um dataframe com 30 observações de 7 variáveis.
\describe{
 \item{\code{pal}}{fator com 10 níveis, palavras-chave do
 experimento.}
 \item{\code{tipo}}{fator com 2 níveis, correspondentes aos tipos de
  vogais, simples (\code{sim}) e dupla (\code{dup})}
 \item{\code{sil}}{fator com 2 níveis, tonicidade da sílaba a que a
 vogal pertence, pré-tônica (\code{pre}) ou tônica (\code{ton}).}
 \item{\code{vog}}{fator com três níveis, identifica a vogal
 analisada.}
 \item{\code{rep}}{fator com três níveis, identifica a repetição da
 sessão de gravação.}
 \item{\code{dur}}{vetor numérico, duração bruta da vogal em
 milissegundos.}
 \item{\code{norm}}{vetor numérico, duração normalizada da vogal em
 z-\emph{scores}.}
}
}
\source{
Dados coletados por Pablo Arantes em 2012 no Laboratório de Fonética
da Universidade Federal de Minas Gerais.
}
\usage{
quantidade
}
\description{
Dados de duração de vogais simples e duplas em pares de palavras como
"catinga"-"caatinga" e "copta"-"coopta" na fala lida do português
brasileiro.
}
\details{
Palavras-chave (coluna \code{pal}) foram inseridas em frases-veículo
do tipo "Digo \emph{palavra-chave} [bem devagar/para ele/lentamente]".
Foram gravadas três repetições de cada frase-veículo em uma única
sessão de gravação. A ordem das frases foi aleatorizada antes de cada
reapresentação da lista de estímulos ao informante. As frases foram
apresentadas uma por vez na forma escrita. A informante é do sexo
feminino, natural de Recife-PE, com 40 anos na época da gravação.
}
\examples{
data(quantidade)
summary(quantidade)
}
\keyword{datasets}
