% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxaeloc}
\alias{taxaeloc}
\title{Duração de sílabas fonéticas}
\format{
Um dataframe com 897 observações de 2 variáveis.
\describe{
 \item{suj}{fator com 5 níveis, correspondentes aos identificadores
 dos informantes femininos.}
 \item{dur}{vetor numérico, duração em ms das sílabas fonéticas de
 trechos da leitura do mesmo texto pelas informantes.}
}
}
\source{
Dados coletados por Plinio A. Barbosa entre 2009 e 2014 no Grupo de
Estudos de Prosódia da Fala da Universidade Estadual de Campinas.
}
\usage{
taxaeloc
}
\description{
Dados de duração em milissegundos (ms) de sílabas fonéticas na fala
lida de cinco mulheres falantes da variedade do português brasileiro
falado na região sudeste do Brasil.
}
\details{
As informantes leram um texto de cerca de 1600 palavras sobre a
origem dos Pastéis de Belém. Trechos aleatórios entre 10 e 20
segundos foram extraídos com a anotação das sílabas fonéticas. Um
script do Praat extraiu as durações dessas mesmas sílabas em
milissegundos, gerando um arquivo de texto que também contém a
identificação da informante e a etiqueta da sílaba (não informada
aqui).

Algumas das sílabas que vão do início de uma vogal ao início da
próxima vogal subsequente, incluem uma pausa silenciosa, tendo em
vista o objetivo de assinalar a taxa de elocução a partir da
distribuição de suas durações.

As informantes são todas do interior de São Paulo, com idade entre 20
e 35 anos no momento da gravação.
}
\examples{
data(taxaeloc)
summary(taxaeloc)
}
\keyword{datasets}
