% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tonisil}
\alias{tonisil}
\title{Sílabas tônicas e átonas em PB}
\format{
Um dataframe com 10764 observações de 8 variáveis.
\describe{
 \item{sil}{vetor de caracteres, representação ortográfica das
 sílabas.}
 \item{ton}{fator com dois níveis, correspondentes aos graus de
 tonicidade das sílabas, tônica (\code{1}) e átona (\code{0}).}
 \item{tipo}{fator com 3 níveis, tipo de entrevista, diálogo entre
 dois informantes (\code{d2}), diálogo entre informante e
 documentador (\code{did}) e elocução formal (\code{ef}).}
 \item{sexo}{fator com 2 níveis, sexo do informante.}
 \item{idade}{vetor numérico, idade do informante no momento da
 gravação.}
 \item{faixa}{fator com três níveis, correspondentes à faixa etária
 do informante, 25--35 anos (\code{1}), 36--25 anos (\code{2}) e mais
 do que 56 anos (\code{3}).}
 \item{ano}{fator com 7 níveis, correspondentes ano em que a
 entrevista foi realizada.}
 \item{cod}{fator com 9 níveis, correspondentes ao código
 identificador da entrevista.}
}
}
\source{
A transcrição das entrevistas em forma ortográfica foi encontrada no
site \url{http://www.letras.ufrj.br/nurc-rj/}.

Francimeire Coelho, Henrique Claro e Silvério Donatti, estudantes do
Bacharelado em Linguística da Universidade Federal de São Carlos
fizeram a seleção manual das palavras em três entrevistas em 2013.

Os autores do sistema Petrus gentilmente processaram a lista de
palavras usando uma versão ainda não pública do sistema.

Pablo Arantes fez correções nos resultados produzidos pelo sistema
Petrus e preparou o dataframe no segundo semestre de 2013.
}
\usage{
tonisil
}
\description{
Dados de frequência de sílabas tônicas e átonas em portugês brasileiro
(PB) em uma amostra do projeto NURC (Projeto da Norma Urbana Oral
Culta).
}
\details{
As entrevistas ou inquéritos coletadas no âmbito do projeto NURC
dividem-se em três tipos:
\itemize{
 \item Elocução formal (EF): transcrição de aulas, coferências,
 palestras.
 \item Diálogos entre informante e documentador (DID): entrevistas
 sobre diferentes temas feitas diretamente entre
 entrevistador/entrevistado.
 \item Diálogo entre dois informantes (D2): gravação de um diálogo
 (conversa) entre dois entrevistados com a presença de um
 documentador.
}

Para este conjunto de dados foram retiradas em torno de 600 palavras
do início de cada entrevista.

Nas entrevistas do tipo DID, as palavras foram retiradas apenas dos
turnos do informante e nas do tipo D2  foram incluídas apenas as do
turno de um dos locutores (sempre aqueles identificados como L1).

As entrevistas gravadas foram transcritas pela equipe do NURC e estão
disponíveis na forma  ortográfica. Além das palavras propriamente
ditas, as transcrições registram graficamente eventos de fala como
hesitações, palavras interrompidas, sobreposições e pausas. As
palavras para cada amostra foram selecionadas manualmente para
excluir sinais de pontuação e outros eventos de fala que não podem
ser considerados vocábulos. A lista de palavras na forma ortográfica
foi submetida ao programa Petrus (Phonetic transcriber for brazilian
user support), acessível a partir de
\url{www.nilc.icmc.usp.br/petrus/}. O programa faz automaticamente a
separação das palavras em sílabas e indica a tonicidade das sílabas.
Essa informação foi usada para a produção do dataframe. Os erros de
separação silábica e atribuição de tonicidade encontrados foram
corrigidos manualmente. Em particular o programa, na versão usada
quando da produção dos dados, não seguia os mesmos critérios para
atribuição de atonicidade para os monossílabos e dissílabos.

A atribuição de atonicidade a monossílabos e dissílabos seguiu os
critérios usualmente apontados por gramáticas tradicionais. Em
especial, consultou-se Bechara (2009). Para maior clareza, elenca-se
abaixo a lista de monossílabos que foram classificados como tônicos e
átonos e a lista de dissílabos átonos:
\itemize{
 \item monossílabos átono: o(s), a(s), um, uns, me, te, se, lhe(s),
 nos, de, do(s), da(s), em, no(s), na(s), e, que.
 \item monossílabos tônicos: substantivos, adjetivos e verbos - ex.:
 mil, trás, pé, só, rei, vã, ri, vá, há.
 \item dissílabos átonos: uma(s), como (conj.), porque (conj.), para
 (prep.), pelo(a).
}
}
\examples{
data(tonisil)
summary(tonisil)
}
\references{
BECHARA, Evanildo. \emph{Moderna gramática portuguesa}. 37. ed. Rio
de Janeiro: Editora Nova Fronteira, 2009.
}
\keyword{datasets}
