% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{upsid}
\alias{upsid}
\title{Número de segmentos fonológicos por língua}
\format{
Um dataframe com 451 observations de duas variáveis.
\describe{
 \item{nseg}{vetor numérico, número de segmentos fonológicos na língua.}
 \item{lang}{fator com 451 níveis, nome da língua.}
}
}
\source{
\url{http://web.phonetik.uni-frankfurt.de/upsid_nr_seg.html}
}
\usage{
upsid
}
\description{
Dados retirados do UCLA Phonological Segment Inventory Database
(UPSID), base de dados organizada por Ian Maddieson e Kristin Precoda
que reúne informações sobre a distribuição de 919 diferentes
segmentos fonológicos em 451 línguas.
}
\details{
Uma interface web simplificada para o UPSID está disponível em
\url{http://web.phonetik.uni-frankfurt.de/upsid.html}. Uma das
consultas pré-compiladas disponiblizadas naquele \emph{site} serviu
de base para a preparação deste conjunto de dados.
}
\examples{
data(upsid)
summary(upsid)
}
\references{
MADDIESON, Ian. \emph{Patterns of sounds}. Cambridge: Cambridge University
Press, 1984.
}
\keyword{datasets}
