% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{word_order}
\alias{word_order}
\title{Ordem dos constituintes sintáticos na frase}
\format{
Um dataframe com 1377 observações de 4 variáveis.
\describe{
  \item{code}{fator com 1377 níveis, código identificador da língua.}
  \item{name}{fator com 1377 níveis, nome da língua.}
  \item{description}{fator com 7 níveis, ordem preferencial dos
  constituintes; Objeto-Sujeito-Verbo (\code{OSV}),
  Objeto-Verbo-Sujeito (\code{OVS}),
  Sujeito-Objeto-Verbo (\code{SOV}),
  Sujeito-Verbo-Objeto (\code{SVO}),
  Verbo-Objeto-Sujeito \code{VOS}, Verbo-Sujeito-Objeto (\code{VSO})
  e sem ordem preferencial (\code{ndo}).}
  \item{family}{fator com 171 níveis, família a que pertence a língua.}
}
}
\source{
\url{http://wals.info/}
}
\usage{
word_order
}
\description{
Ordem dos constituintes sintáticos sujeito, verbo e objeto na frase
em uma amostra de 1377 línguas vivas.
}
\details{
Os dados foram compilados pelos autores do The World Atlas of
Language Structures (WALS). O WALS é descrito por seus organizadores
como uma base da dados de propriedades linguísticas estruturais
coletadas em trabalhos descritivos.
}
\examples{
data(word_order)
summary(word_order)
}
\references{
DRYER, Matthew S.; HASPELMATH, Martin (Orgs.). \emph{WALS Online}.
Leipzig: Max Planck Institute for Evolutionary Anthropology, 2013.
Disponível em: \url{https://wals.info/}.
}
\keyword{datasets}
